dependency refactoring and code reformating done
