+++ src/model/common/imageModels.ts
